cloud_function/requirements.txt
sqlalchemy
psycopg2-binary
requests
pandas
openai
python-dotenv
functions-framework